Rename datatype header source to dataType
