+++ apx/common/src/apx_dataSignature.c
@@ -40,7 +40,7 @@
 #include "adt_hash.h"
 #include "adt_str.h"
 #include "apx_types.h"
-#include "apx_datatype.h"
+#include "apx_dataType.h"
 #ifdef MEM_LEAK_CHECK
 #include "CMemLeak.h"
 #endif
