+++ apx/common/src/apx_dataType.c
@@ -31,7 +31,7 @@
 #include <errno.h>
 #include <assert.h>
 #include <malloc.h>
-#include "apx_datatype.h"
+#include "apx_dataType.h"
 #include "apx_types.h"
 #ifdef MEM_LEAK_CHECK
 #include "CMemLeak.h"
