+++ apx/common/src/apx_node.c
@@ -11,7 +11,7 @@
 #include "apx_types.h"
 #include "adt_ary.h"
 #include "adt_hash.h"
-#include "apx_datatype.h"
+#include "apx_dataType.h"
 #ifdef MEM_LEAK_CHECK
 #include "CMemLeak.h"
 #endif
